refactor: 바텀시트 slot 구조로 변경
- 바텀시트 위치 변경
- 바텀시트 버튼 props로 변경
- 빈 input 입력 시 alert 처리
